+++ tests/com/tkomiya/testutils/TestVocabListLoader.java
@@ -14,6 +14,7 @@ import com.tkomiya.main.Translation;
 import com.tkomiya.models.Vocab;
 import com.tkomiya.models.Vocab.SupportedLanguage;
 import com.tkomiya.models.VocabList;
+import com.tkomiya.testutils.datageneration.JapaneseTranslationFactory;
 import com.tkomiya.testutils.datageneration.KoreanTranslationFactory;
 import com.tkomiya.testutils.datageneration.EnglishTranslationFactory;
 
@@ -41,15 +42,15 @@ public class TestVocabListLoader {
 	}
 	
 	public static Vocab mondayVocab() {
-	  return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.mondayEnglishTranslation(), KoreanTranslationFactory.mondayKoreanTranslation(), TestTranslationLoader.mondayJapaneseTranslation());
+	  return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.mondayEnglishTranslation(), KoreanTranslationFactory.mondayKoreanTranslation(), JapaneseTranslationFactory.mondayJapaneseTranslation());
 	}
 	
 	public static Vocab tuesdayVocab() {
-	  return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.tuesdayEnglishTranslation(), KoreanTranslationFactory.tuesdayKoreanTranslation(), TestTranslationLoader.tuesdayJapaneseTranslation());
+	  return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.tuesdayEnglishTranslation(), KoreanTranslationFactory.tuesdayKoreanTranslation(), JapaneseTranslationFactory.tuesdayJapaneseTranslation());
 	}
 	
 	public static Vocab wednesdayVocab() {
-		return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.wednesdayEnglishTranslation(), KoreanTranslationFactory.wednesdayKoreanTranslation(), TestTranslationLoader.wednesdayJapaneseTranslation());
+		return makeVocab(SupportedLanguage.ENGLISH, EnglishTranslationFactory.wednesdayEnglishTranslation(), KoreanTranslationFactory.wednesdayKoreanTranslation(), JapaneseTranslationFactory.wednesdayJapaneseTranslation());
 	}
 	
 	private static Vocab makeVocab(SupportedLanguage primaryLanguage, Translation englishTranslation, Translation koreanTranslation, Translation japaneseTranslation) {
